prism_intern/requirements.txt
asgiref==3.4.1
distlib==0.3.4
Django==4.0.1
djongo==1.3.6
dnspython==2.2.0
mongoengine==0.23.1
Pillow==9.0.0
psycopg2==2.9.3
pymongo==3.12.3
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
urllib3==1.26.8
